Add a big patch for the Mingus framework.
This project relies on Mingus, but mingus is as yet incomplete.

This project relies on the features added/fixed by applying this patch
to mingus.

This patch applies to the last patch in the official Mingus repo
dated November 27, 2009.

Features and Fixes include:

Adds a bunch of functionality and helper functions.
- Fixes lilypond, png, pdf and midi exporting.

Fixes midi importing:
- Previously, could only deal with homophony. Should be able to handle
polyphony now (untested. not sure if Mingus Bar objects even support polyphony)
- Previously, each note was given the duration that should've applied to the
previous note. Fixed (tested).
- Cavevat: rhythms drift somehow, during the import of some multi-voice
and rhythmically complex midi files.

Rewrite midi import to use quantized ints to keep track of the timing.

Add new method to bars so that rests are properly created when reading in midi
files.
- Fixes typesetting problem when rests aren't explicitly added in lilypond.

Explicitly add bar lines to tracks, and don't print a track if it has no notes
in it.
- Cleans up Lilypond printouts.